Fixed library files structure
