+++ athena_udf/utils.py
@@ -7,4 +7,4 @@ T = TypeVar("T")
 def get_chunks(lst: list[T], n: int) -> Generator[list[T], None, None]:
     """Yield successive n-sized chunks from lst."""
     for i in range(0, len(lst), n):
-        yield lst[i : i + n]
+        yield lst[i: i + n]
